USB/ANT/FORTIUS: Stricter error checking
Return an error if either usb_set_configuration() or
usb_claim_interface() fail. This should help in the case
where the device has insufficient permissions (Linux) or
is already in use by ANT Agent (Windows/MAC).

Note that Linux support is complicated by the fact that we
have to forcibly disconnect the usb_serial driver. Because
of this, we attempt to grab the first device regardless of
whether it was already in use. For other OS, this patch
should enable the use of multiple ANT sticks (for instance,
one for Zwift and one for GC).